+++ src/ado2gh/Commands/IntegrateBoards/IntegrateBoardsCommandArgs.cs
@@ -8,9 +8,8 @@ namespace OctoshiftCLI.AdoToGithub.Commands.IntegrateBoards
         public string AdoTeamProject { get; set; }
         public string GithubOrg { get; set; }
         public string GithubRepo { get; set; }
+        public string ServiceConnectionId { get; set; }
         [Secret]
         public string AdoPat { get; set; }
-        [Secret]
-        public string GithubPat { get; set; }
     }
 }